version 2

+++ BalisticCalculator/_windage.c
@@ -1,19 +1,16 @@
-double Windage(double WindSpeed, double Vi, double xx, double t){
-	double Vw = WindSpeed*17.60; // Convert to inches per second.
-	return (Vw*(t-xx/Vi));
+float Windage(float WindSpeed, float Vi, float xx, float t) {
+	float Vw = WindSpeed * 17.60; // Convert to inches per second.
+	return (Vw * (t - xx / Vi));
 }
 
-
 // Headwind is positive at WindAngle=0
-double HeadWind(double WindSpeed, double WindAngle){
-	double Wangle = DegtoRad(WindAngle);
-	return (cos(Wangle)*WindSpeed);
+float HeadWind(float WindSpeed, float WindAngle) {
+	float Wangle = DegtoRad(WindAngle);
+	return (cos(Wangle) * WindSpeed);
 }
 
 // Positive is from Shooter's Right to Left (Wind from 90 degree)
-double CrossWind(double WindSpeed, double WindAngle){
-	double Wangle = DegtoRad(WindAngle);
-	return (sin(Wangle)*WindSpeed);
+float CrossWind(float WindSpeed, float WindAngle) {
+	float Wangle = DegtoRad(WindAngle);
+	return (sin(Wangle) * WindSpeed);
 }
-
-
